pyproject.toml
[tool.poetry]
name = "SRGAN"
version = "0.1.0"
description = ""
authors = ["Yudai Hayashi <yudai0913haya@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">3.9,<3.11"
tensorflow = "^2.10.0"
Pillow = "^9.2.0"
matplotlib = "^3.6.1"
tqdm = "^4.64.1"
PyYAML = "^6.0"
scikit-learn = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
